+++ src/convtools/__init__.py
@@ -1,4 +1,5 @@
 """Public API."""
+
 from ._conversion import conversion  # noqa: F401
 from ._dt import DateGrid, DateTimeGrid