+++ src/convtools/_heuristics.py
@@ -1,4 +1,5 @@
 """Helpers to collect info about environment."""
+
 import sys