+++ src/convtools/_utils.py
@@ -4,6 +4,7 @@ e.g.
  - recently used cache
  - options ctx manager
 """
+
 import os
 import sys
 import tempfile